requirements.txt
# BlockHire Production Dependencies
# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0

# CORS and Security
django-cors-headers==4.3.1

# Authentication
PyJWT==2.8.0

# Image Processing
Pillow==10.4.0

# Configuration
python-decouple==3.8

# Database
psycopg[binary]==3.2.10
dj-database-url==2.1.0

# Static Files
whitenoise==6.6.0

# Production Server
gunicorn==21.2.0

# API Features
django-filter==23.3
django-ratelimit==4.1.0
django-cleanup==9.0.0

# Cloud Storage
boto3==1.34.0
django-storages==1.14.2